Minor edits, and bug note
My edits: I added a bit of specificity regarding which chronology categories need to be run for the specific examples we have provided.

Also note - there is an error on line 161. The object "alldates" was not found.

At the top of this script, you have a note about "alldates" that refers to the section "Add to main date table", which said this is optional. But, "alldates" appears first in Code Chunk 10, in the "Assign types" within "Compile chronologies" section. Because this code chunk creates "typeindex", I can't complete the remainder of this section